More structural changes and caching work
