+++ client/flutter/fadzmaq/lib/controllers/globalData.dart
@@ -1,6 +1,8 @@
+import 'package:fadzmaq/controllers/cache.dart';
 import 'package:fadzmaq/controllers/request.dart';
 import 'package:fadzmaq/models/app_config.dart';
 import 'package:fadzmaq/models/globalModel.dart';
+import 'package:fadzmaq/models/hobbies.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:fadzmaq/controllers/globals.dart';
 import 'package:fadzmaq/models/matches.dart';
@@ -9,6 +11,7 @@ import 'package:fadzmaq/models/recommendations.dart';
 import 'package:fadzmaq/views/loginscreen.dart';
 import 'package:flutter/material.dart';
 import 'dart:convert';
+import 'package:http/http.dart' as http;
 
 class GlobalData extends InheritedWidget {
   GlobalData({
@@ -23,7 +26,11 @@ class GlobalData extends InheritedWidget {
         .model;
   }
 
-  // TODO this should change I think...
+  // This widget is created once and does not change, therefore has no
+  // concept of notifying widgets below it of a change
+  // as widgets are rebuilt they will use whatever changes are present
+  // in the model, but live changes must be managed by that widget's
+  // own state management
   @override
   bool updateShouldNotify(InheritedWidget oldWidget) => false;
 }
@@ -49,7 +56,8 @@ class _VerifyModelState extends State<VerifyModel> {
   @override
   void didChangeDependencies() {
     if (_future == null) {
-      String server = AppConfig.of(context).server + _getURL(widget.model);
+      String server =
+          AppConfig.of(context).server + _getURL(widget.model);
       _future = httpGet(server);
     }
 
@@ -58,7 +66,7 @@ class _VerifyModelState extends State<VerifyModel> {
 
   @override
   Widget build(BuildContext context) {
-    if (_checkModel(widget.model)) {
+    if (_checkModel(context, widget.model)) {
       return widget.builder(context);
     }
 
@@ -70,7 +78,7 @@ class _VerifyModelState extends State<VerifyModel> {
             return Center(child: Text(snapshot.data.toString()));
           }
           if (snapshot.data.statusCode == 200) {
-            _loadJSON(widget.model, json.decode(snapshot.data.body));
+            _loadJSON(context, widget.model, json.decode(snapshot.data.body));
             return widget.builder(context);
           } else if (snapshot.data.statusCode == 401) {
             // not sure if this is the best way to do this but it works for now - Jordan
@@ -93,55 +101,108 @@ class _VerifyModelState extends State<VerifyModel> {
       },
     );
   }
+}
 
-  String _getURL(Model model) {
-    String url;
-    switch (widget.model) {
-      case Model.matches:
-        url = Globals.matchesURL;
-        break;
-      case Model.recommendations:
-        url = Globals.recsURL;
-        break;
-      case Model.userProfile:
-        url = Globals.profileURL;
-        break;
-    }
+Future loadGlobalModels(BuildContext context) async {
+  List<Future> futures = List<Future>();
 
-    return url;
+  for (Model model in Model.values) {
+    futures.add(httpLoadModel(context, model));
   }
 
-  bool _checkModel(Model model) {
-    GlobalModel mainModel = GlobalData.of(context);
-    bool present;
-    switch (widget.model) {
-      case Model.matches:
-        present = mainModel.matches != null;
-        break;
-      case Model.recommendations:
-        present = mainModel.recommendations != null;
-        break;
-      case Model.userProfile:
-        present = mainModel.userProfile != null;
-        break;
-    }
-    return present;
+  await Future.wait(futures);
+}
+
+/// This is pretty hacky but it lets the future for the request stay
+/// umcompleted until the images are also cached.
+Future httpLoadModel(BuildContext context, Model model) async {
+  String server = AppConfig.of(context).server + _getURL(model);
+  http.Response response;
+  dynamic responseJson;
+
+  try {
+    response = await httpGet(server);
+    responseJson = json.decode(response.body);
+    _loadJSON(context, model, json.decode(responseJson));
+    _cacheModelImages(context, model);
+  } catch (e) {
+    throw e;
+  }
+}
+
+String _getURL(Model model) {
+  String url;
+  switch (model) {
+    case Model.matches:
+      url = Globals.matchesURL;
+      break;
+    case Model.recommendations:
+      url = Globals.recsURL;
+      break;
+    case Model.userProfile:
+      url = Globals.profileURL;
+      break;
+    case Model.allHobbies:
+      url = Globals.allHobbiesURL;
+      break;
   }
 
-  void _loadJSON(Model model, dynamic json) {
-    GlobalModel mainModel = GlobalData.of(context);
-    switch (widget.model) {
-      case Model.matches:
-        mainModel.matches = MatchesData.fromJson(json);
-        break;
-      case Model.recommendations:
-        mainModel.recommendations = RecommendationsData.fromJson(json);
-        break;
-      case Model.userProfile:
-        var pc = ProfileContainer.fromJson(json);
-        if (pc == null) throw Exception("empty Profile Container");
-        mainModel.userProfile = pc.profile;
-        break;
-    }
+  return url;
+}
+
+bool _checkModel(BuildContext context, Model model) {
+  GlobalModel mainModel = GlobalData.of(context);
+  bool present;
+  switch (model) {
+    case Model.matches:
+      present = mainModel.matches != null;
+      break;
+    case Model.recommendations:
+      present = mainModel.recommendations != null;
+      break;
+    case Model.userProfile:
+      present = mainModel.userProfile != null;
+      break;
+    case Model.allHobbies:
+      present = mainModel.allHobbies != null;
+      break;
+  }
+  return present;
+}
+
+void _loadJSON(BuildContext context, Model model, dynamic json) {
+  GlobalModel mainModel = GlobalData.of(context);
+  switch (model) {
+    case Model.matches:
+      mainModel.matches = MatchesData.fromJson(json);
+      break;
+    case Model.recommendations:
+      mainModel.recommendations = RecommendationsData.fromJson(json);
+      break;
+    case Model.userProfile:
+      var pc = ProfileContainer.fromJson(json);
+      if (pc == null) throw Exception("empty Profile Container");
+      mainModel.userProfile = pc.profile;
+      break;
+    case Model.allHobbies:
+      mainModel.allHobbies = AllHobbiesData.fromJson(json);
+      break;
+  }
+}
+
+_cacheModelImages(BuildContext context, Model model) {
+  GlobalModel globalModel = GlobalData.of(context);
+  switch (model) {
+    case Model.matches:
+      cacheMatchPhotos(context, globalModel.matches);
+      break;
+    case Model.recommendations:
+      cacheRecommendationPhotos(context, globalModel.recommendations);
+      break;
+    case Model.userProfile:
+      cacheProfilePhotos(context, globalModel.userProfile);
+      break;
+    case Model.allHobbies:
+      break;
   }
 }
