+++ client/flutter/fadzmaq/lib/controllers/globals.dart
@@ -7,6 +7,7 @@ class Globals{
   static const String profileURL = "profile";
   static const String matchesURL = "matches";
   static const String recsURL = "user/recs";
+  static const String allHobbiesURL = "hobbies";
